requirements/local.txt
-r base.txt

django-extensions==2.0.7
six==1.11.0